requirements.txt
beautifulsoup4==4.10.0
python-telegram-bot==21.0.1
Requests==2.31.0
